+++ common-lib/src/main/kotlin/com/example/banking/common/dto/TransactionDto.kt
@@ -3,6 +3,7 @@ package com.example.banking.common.dto
 import jakarta.validation.constraints.DecimalMin
 import jakarta.validation.constraints.NotNull
 import jakarta.validation.constraints.Positive
+import java.math.BigDecimal
 
 //@Schema(description = "Transaction request")
 data class TransactionDTO(
@@ -12,10 +13,10 @@ data class TransactionDTO(
     val receiverId: Long,
     @field:Positive(message = "Amount must be positive")
     @field:DecimalMin("0.01")
-    val amount: Double
+    val amount: BigDecimal
 ) {
     init {
         require(senderId != receiverId) { "Sender and receiver must be different" }
-        require(amount > 0) { "Amount must be positive" }
+        require(amount > BigDecimal(0)) { "Amount must be positive" }
     }
 }
